backend/requirements.txt
fastapi==0.97.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pyjwt==2.8.0
passlib[bcrypt]==1.7.5
scikit-learn==1.3.2
numpy==1.26.4
pydantic==1.10.11
requests==2.32.4
sqlalchemy==2.0.0
slowapi==0.1.8
python-dotenv==1.0.0
